Fixed a series more preprocessor and type issues shown in the windows build

+++ libAudio/string.hxx
@@ -3,8 +3,8 @@
 
 #include "uniquePtr.hxx"
 
-inline uint32_t stringsLength(const char *const part) noexcept { return strlen(part); }
-template<typename... Args> inline uint32_t stringsLength(const char *const part, Args &&...args) noexcept
+inline size_t stringsLength(const char *const part) noexcept { return strlen(part); }
+template<typename... Args> inline size_t stringsLength(const char *const part, Args &&...args) noexcept
 	{ return strlen(part) + stringsLength(args...); }
 
 inline  std::unique_ptr<char []> stringsConcat(std::unique_ptr<char []> &&dst, const uint32_t offset, const char *const part) noexcept
